fix JSON loading error
JSON matrices weren't loading when the package was pip installed, only
when the entire source code was present. Changed build to hatchling to
include json files without the addition of another file format like a
MANIFEST.in. Added tests to ensure that files were being correctly
loaded using this new paradigm but also ensured that files can still be
locally loaded parimarily in the case of a dev environment.